@@ -9,29 +9,39 @@
     }
   ],
   "require": {
-    "php": "^7.1",
-    "ext-curl": "*",
+    "php": "^8.1",
     "ext-xml": "*",
-    "guzzlehttp/guzzle": "~6.3",
-    "psr/log": "~1.0",
-    "symfony/serializer": "~3.4"
+    "psr/http-client-implementation": "^1.0",
+    "psr/http-factory-implementation": "^1.0",
+    "psr/http-message-implementation": "^1.0",
+    "psr/log": "^3.0",
+    "symfony/serializer": "^6.4"
   },
   "require-dev": {
-    "drupal/coder": ">=8.2.12",
-    "phpunit/phpunit": "~6.5",
-    "symfony/phpunit-bridge": "~3.4"
+    "dealerdirect/phpcodesniffer-composer-installer": "^1.0",
+    "guzzlehttp/psr7": "^2.6",
+    "http-interop/http-factory-guzzle": "^1.2",
+    "php-http/mock-client": "^1.0",
+    "phpstan/phpstan": "^1.10",
+    "phpunit/phpunit": "^9.6",
+    "previousnext/coding-standard": "^0.1.3",
+    "staabm/annotate-pull-request-from-checkstyle": "^1.8"
   },
   "autoload": {
     "psr-4": {"BomWeather\\": "src/"}
   },
   "autoload-dev": {
-    "psr-4": {"BomWeather\\Tests\\": "tests/"}
+    "psr-4": {"BomWeather\\Tests\\": "tests/src"}
   },
   "config": {
     "bin-dir": "bin/",
     "preferred-install": {
       "*": "dist"
     },
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "dealerdirect/phpcodesniffer-composer-installer": true,
+      "php-http/discovery": true
+    }
   }
 }
