@@ -1,177 +1,130 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace BomWeather\Forecast;
 
 /**
  * A value object for forecast periods.
  */
-class ForecastPeriod {
+final class ForecastPeriod {
 
   /**
    * The start time.
-   *
-   * @var \DateTime
    */
-  protected $startTime;
+  protected ?\DateTimeImmutable $startTime = NULL;
 
   /**
    * The end time.
-   *
-   * @var \DateTime
    */
-  protected $endTime;
+  protected ?\DateTimeImmutable $endTime = NULL;
 
   /**
    * The air temp maximum in Celsius.
-   *
-   * @var int
    */
-  protected $airTempMaximum;
+  protected ?int $airTempMaximum = NULL;
 
   /**
    * The probability of precipitation.
-   *
-   * @var string
    */
-  protected $probabilityOfPrecipitation;
+  protected ?string $probabilityOfPrecipitation = NULL;
 
   /**
    * The precipitation range.
-   *
-   * @var string
    */
-  protected $precipitationRange;
+  protected ?string $precipitationRange = NULL;
 
   /**
    * The forecast icon code.
-   *
-   * @var string
    */
-  protected $iconCode;
+  protected ?string $iconCode = NULL;
 
   /**
    * The air temp minimum in Celsius.
-   *
-   * @var int
    */
-  protected $airTempMinimum;
+  protected ?int $airTempMinimum = NULL;
 
   /**
    * The precis.
-   *
-   * @var string
    */
-  protected $precis;
+  protected ?string $precis = NULL;
 
   /**
    * The UV alert.
-   *
-   * @var string
    */
-  protected $uvAlert;
+  protected ?string $uvAlert = NULL;
 
   /**
    * The forecast.
-   *
-   * @var string
    */
-  protected $forecast;
+  protected ?string $forecast = NULL;
 
   /**
    * The synoptic situation.
-   *
-   * @var string
    */
-  protected $synoptic;
+  protected ?string $synoptic = NULL;
 
   /**
    * The seas.
-   *
-   * @var string
    */
-  protected $seas;
+  protected ?string $seas = NULL;
 
   /**
    * The swell.
-   *
-   * @var string
    */
-  protected $swell;
+  protected ?string $swell = NULL;
 
   /**
    * The weather.
-   *
-   * @var string
    */
-  protected $weather;
+  protected ?string $weather = NULL;
 
   /**
    * The winds.
-   *
-   * @var string
    */
-  protected $winds;
+  protected ?string $winds = NULL;
 
   /**
    * Factory method.
    *
    * @return $this
    */
-  public static function create() {
+  public static function create(): static {
     return new static();
   }
 
   /**
    * Gets the start time.
-   *
-   * @return \DateTime
-   *   The start time.
    */
-  public function getStartTime(): \DateTime {
+  public function getStartTime(): ?\DateTimeImmutable {
     return $this->startTime;
   }
 
   /**
    * Sets the start time.
-   *
-   * @param \DateTime $startTime
-   *   The start time.
-   *
-   * @return $this
    */
-  public function setStartTime(\DateTime $startTime): ForecastPeriod {
+  public function setStartTime(\DateTimeImmutable $startTime): ForecastPeriod {
     $this->startTime = $startTime;
     return $this;
   }
 
   /**
    * Gets the end time.
-   *
-   * @return \DateTime
-   *   The end time.
    */
-  public function getEndTime(): \DateTime {
+  public function getEndTime(): ?\DateTimeImmutable {
     return $this->endTime;
   }
 
   /**
    * Sets the end time.
-   *
-   * @param \DateTime $endTime
-   *   The end time.
-   *
-   * @return $this
    */
-  public function setEndTime(\DateTime $endTime): ForecastPeriod {
+  public function setEndTime(\DateTimeImmutable $endTime): ForecastPeriod {
     $this->endTime = $endTime;
     return $this;
   }
 
   /**
    * Gets the icon code.
-   *
-   * @return string
-   *   The icon code.
    */
   public function getIconCode(): ?string {
     return $this->iconCode;
@@ -179,11 +132,6 @@ class ForecastPeriod {
 
   /**
    * Sets the icon code.
-   *
-   * @param string $iconCode
-   *   The icon code.
-   *
-   * @return $this
    */
   public function setIconCode(string $iconCode): ForecastPeriod {
     $this->iconCode = $iconCode;
@@ -192,9 +140,6 @@ class ForecastPeriod {
 
   /**
    * Gets the precipitation range.
-   *
-   * @return string
-   *   The precipitation range.
    */
   public function getPrecipitationRange(): ?string {
     return $this->precipitationRange;
@@ -202,23 +147,14 @@ class ForecastPeriod {
 
   /**
    * Sets the precipitation range.
-   *
-   * @param string $precipitationRange
-   *   The precipitation range.
-   *
-   * @return $this
    */
-  public function setPrecipitationRange(string $precipitationRange
-  ): ForecastPeriod {
+  public function setPrecipitationRange(string $precipitationRange): ForecastPeriod {
     $this->precipitationRange = $precipitationRange;
     return $this;
   }
 
   /**
    * Gets the air temp minimum.
-   *
-   * @return int
-   *   The air temp minimum.
    */
   public function getAirTempMinimum(): ?int {
     return $this->airTempMinimum;
@@ -226,11 +162,6 @@ class ForecastPeriod {
 
   /**
    * Sets the air temp minimum.
-   *
-   * @param int $airTempMinimum
-   *   The air temp minimum.
-   *
-   * @return $this
    */
   public function setAirTempMinimum(int $airTempMinimum): ForecastPeriod {
     $this->airTempMinimum = $airTempMinimum;
@@ -239,9 +170,6 @@ class ForecastPeriod {
 
   /**
    * Gets the air temp maximum.
-   *
-   * @return int
-   *   The air temp maximum.
    */
   public function getAirTempMaximum(): ?int {
     return $this->airTempMaximum;
@@ -249,23 +177,14 @@ class ForecastPeriod {
 
   /**
    * Sets the air temp maximum.
-   *
-   * @param int $airTempMaximum
-   *   The air temp maximum.
-   *
-   * @return $this
    */
-  public function setAirTempMaximum(int $airTempMaximum
-  ): ForecastPeriod {
+  public function setAirTempMaximum(int $airTempMaximum): ForecastPeriod {
     $this->airTempMaximum = $airTempMaximum;
     return $this;
   }
 
   /**
    * Gets the precis.
-   *
-   * @return string
-   *   The precis.
    */
   public function getPrecis(): ?string {
     return $this->precis;
@@ -273,11 +192,6 @@ class ForecastPeriod {
 
   /**
    * Sets the precis.
-   *
-   * @param string $precis
-   *   The precis.
-   *
-   * @return $this
    */
   public function setPrecis(string $precis): ForecastPeriod {
     $this->precis = $precis;
@@ -286,9 +200,6 @@ class ForecastPeriod {
 
   /**
    * Gets the probability of precipitation.
-   *
-   * @return string
-   *   The probability of precipitation.
    */
   public function getProbabilityOfPrecipitation(): ?string {
     return $this->probabilityOfPrecipitation;
@@ -296,11 +207,6 @@ class ForecastPeriod {
 
   /**
    * Sets the probability of precipitation.
-   *
-   * @param string $probabilityOfPrecipitation
-   *   The probability of precipitation.
-   *
-   * @return $this
    */
   public function setProbabilityOfPrecipitation(
     string $probabilityOfPrecipitation
@@ -311,9 +217,6 @@ class ForecastPeriod {
 
   /**
    * Gets the forecast.
-   *
-   * @return string
-   *   The forecast.
    */
   public function getForecast(): ?string {
     return $this->forecast;
@@ -321,11 +224,6 @@ class ForecastPeriod {
 
   /**
    * Sets the forecast.
-   *
-   * @param string $forecast
-   *   The forecast.
-   *
-   * @return $this
    */
   public function setForecast(string $forecast): ForecastPeriod {
     $this->forecast = $forecast;
@@ -334,9 +232,6 @@ class ForecastPeriod {
 
   /**
    * Gets the UV alert.
-   *
-   * @return string
-   *   The UV alert.
    */
   public function getUvAlert(): ?string {
     return $this->uvAlert;
@@ -344,11 +239,6 @@ class ForecastPeriod {
 
   /**
    * Sets the UV alert.
-   *
-   * @param string $uvAlert
-   *   The UV alert.
-   *
-   * @return $this
    */
   public function setUvAlert(string $uvAlert): ForecastPeriod {
     $this->uvAlert = $uvAlert;
@@ -357,21 +247,13 @@ class ForecastPeriod {
 
   /**
    * Gets the synoptic.
-   *
-   * @return string
-   *   The synoptic.
    */
-  public function getSynoptic(): string {
+  public function getSynoptic(): ?string {
     return $this->synoptic;
   }
 
   /**
    * Sets the synoptic.
-   *
-   * @param string $synoptic
-   *   The synoptic.
-   *
-   * @return $this
    */
   public function setSynoptic(string $synoptic): ForecastPeriod {
     $this->synoptic = $synoptic;
@@ -380,21 +262,13 @@ class ForecastPeriod {
 
   /**
    * Gets the seas.
-   *
-   * @return string
-   *   The seas.
    */
-  public function getSeas(): string {
+  public function getSeas(): ?string {
     return $this->seas;
   }
 
   /**
    * Sets the seas.
-   *
-   * @param string $seas
-   *   The seas.
-   *
-   * @return $this
    */
   public function setSeas(string $seas): ForecastPeriod {
     $this->seas = $seas;
@@ -403,21 +277,13 @@ class ForecastPeriod {
 
   /**
    * Gets the swell.
-   *
-   * @return string
-   *   The swell.
    */
-  public function getSwell(): string {
+  public function getSwell(): ?string {
     return $this->swell;
   }
 
   /**
    * Sets the swell.
-   *
-   * @param string $swell
-   *   The swell.
-   *
-   * @return $this
    */
   public function setSwell(string $swell): ForecastPeriod {
     $this->swell = $swell;
@@ -426,21 +292,13 @@ class ForecastPeriod {
 
   /**
    * Gets the weather.
-   *
-   * @return string
-   *   The weather.
    */
-  public function getWeather(): string {
+  public function getWeather(): ?string {
     return $this->weather;
   }
 
   /**
    * Sets the weather.
-   *
-   * @param string $weather
-   *   The weather.
-   *
-   * @return $this
    */
   public function setWeather(string $weather): ForecastPeriod {
     $this->weather = $weather;
@@ -449,21 +307,13 @@ class ForecastPeriod {
 
   /**
    * Gets the winds.
-   *
-   * @return string
-   *   The winds.
    */
-  public function getWinds(): string {
+  public function getWinds(): ?string {
     return $this->winds;
   }
 
   /**
    * Sets the winds.
-   *
-   * @param string $winds
-   *   The winds.
-   *
-   * @return $this
    */
   public function setWinds(string $winds): ForecastPeriod {
     $this->winds = $winds;
