@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace BomWeather\Observation\Serializer;
 
 use BomWeather\Observation\Observation;
@@ -12,13 +14,12 @@ use Symfony\Component\Serializer\Normalizer\DenormalizerInterface;
  */
 class ObservationListNormalizer extends BaseNormalizer {
 
-  protected $supportedInterfaceOrClass = ObservationList::class;
+  protected string|array $supportedInterfaceOrClass = ObservationList::class;
 
   /**
    * {@inheritdoc}
    */
   public function denormalize($data, $class, $format = NULL, array $context = []) {
-
     if (!$this->serializer instanceof DenormalizerInterface) {
       throw new \RuntimeException('The serializer must implement the DenormalizerInterface.');
     }
@@ -26,7 +27,7 @@ class ObservationListNormalizer extends BaseNormalizer {
     $observationList = new ObservationList();
     $observationList->setRefreshMessage($data['observations']['header'][0]['refresh_message']);
 
-    array_map(function ($observationData) use ($observationList) {
+    \array_map(function ($observationData) use ($observationList): void {
       $observationList->addObservation($this->serializer->denormalize($observationData, Observation::class));
     }, $data['observations']['data'], [$observationList]);
 
