@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace BomWeather\Util;
 
 /**
@@ -9,38 +11,25 @@ trait ClassNameSupportNormalizerTrait {
 
   /**
    * The interface or class that this Normalizer supports.
-   *
-   * @var string|array
    */
-  protected $supportedInterfaceOrClass;
+  protected string|array $supportedInterfaceOrClass;
 
   /**
    * The format this Normalizer supports.
-   *
-   * @var string|array
    */
-  protected $format;
+  protected string|array $format;
 
   /**
    * Gets the string or array of supported classes.
-   *
-   * @return array|string
-   *   The string or array of supported classes.
    */
-  public function getSupportedInterfaceOrClass() {
+  public function getSupportedInterfaceOrClass(): string|array {
     return $this->supportedInterfaceOrClass;
   }
 
   /**
    * Sets the string or array of supported classes.
-   *
-   * @param array|string $supported_interface_or_class
-   *   The string or array of supported classes.
-   *
-   * @return $this
-   *   The current object.
    */
-  public function setSupportedInterfaceOrClass($supported_interface_or_class) {
+  public function setSupportedInterfaceOrClass(string|array $supported_interface_or_class): self {
     $this->supportedInterfaceOrClass = $supported_interface_or_class;
     return $this;
   }
@@ -48,16 +37,16 @@ trait ClassNameSupportNormalizerTrait {
   /**
    * {@inheritdoc}
    */
-  public function supportsNormalization($data, $format = NULL) {
+  public function supportsNormalization(mixed $data, string $format = NULL /* , array $context = [] */): bool {
     // If we aren't dealing with an object or the format is not supported return
     // now.
-    if (!is_object($data) || !$this->checkFormat($format)) {
+    if (!\is_object($data) || !$this->checkFormat($format)) {
       return FALSE;
     }
 
     $supported = (array) $this->supportedInterfaceOrClass;
 
-    return (bool) array_filter($supported, function ($name) use ($data) {
+    return (bool) \array_filter($supported, function ($name) use ($data) {
       return $data instanceof $name;
     });
   }
@@ -65,7 +54,7 @@ trait ClassNameSupportNormalizerTrait {
   /**
    * {@inheritdoc}
    */
-  public function supportsDenormalization($data, $type, $format = NULL) {
+  public function supportsDenormalization(mixed $data, string $type, string $format = NULL /* , array $context = [] */): bool {
     // If the format is not supported return now.
     if (!$this->checkFormat($format)) {
       return FALSE;
@@ -74,30 +63,27 @@ trait ClassNameSupportNormalizerTrait {
     $supported = (array) $this->supportedInterfaceOrClass;
 
     $subclass_check = function ($name) use ($type) {
-      return (class_exists($name) || interface_exists($name)) && is_subclass_of($type,
+      return (\class_exists($name) || \interface_exists($name)) && \is_subclass_of($type,
           $name, TRUE);
     };
 
-    return in_array($type, $supported) || array_filter($supported,
+    return \in_array($type, $supported) || \array_filter($supported,
         $subclass_check);
   }
 
   /**
    * Checks if the provided format is supported by this serializer.
    *
-   * @param string $format
-   *   The format to check.
-   *
    * @return bool
    *   TRUE if the format is supported, FALSE otherwise. If no format is
    *   specified this will return TRUE.
    */
-  protected function checkFormat($format = NULL) {
+  protected function checkFormat(string $format = NULL): bool {
     if (!isset($format) || !isset($this->format)) {
       return TRUE;
     }
 
-    return in_array($format, (array) $this->format);
+    return \in_array($format, (array) $this->format);
   }
 
 }
