@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types = 1);
+
 namespace BomWeather\Util;
 
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
@@ -12,6 +14,13 @@ abstract class BaseNormalizer extends AbstractNormalizer {
   use ClassNameSupportNormalizerTrait;
 
   /**
+   * {@inheritdoc}
+   */
+  public function normalize(mixed $object, string $format = NULL, array $context = []) {
+    throw new \RuntimeException("Method not implemented.");
+  }
+
+  /**
    * Checks whether the array is associative.
    *
    * @param array $array
@@ -20,15 +29,8 @@ abstract class BaseNormalizer extends AbstractNormalizer {
    * @return bool
    *   TRUE if the array is associative.
    */
-  protected function isAssoc(array $array) {
-    return (bool) count(array_filter(array_keys($array), 'is_string'));
-  }
-
-  /**
-   * {@inheritdoc}
-   */
-  public function normalize($object, $format = NULL, array $context = []) {
-    throw new \RuntimeException("Method not implemented.");
+  protected function isAssoc(array $array): bool {
+    return (bool) \count(\array_filter(\array_keys($array), 'is_string'));
   }
 
 }
